requirements.txt
PyQt6>=6.4.0
pynput>=1.7.6
mss>=9.0.1
Pillow>=10.0.0
fpdf2>=2.7.0
python-docx>=1.1.0
xlwings>=0.30.0
